refactored static scheduling file writing
