+++ src/main/java/org/cicirello/search/problems/scheduling/WeightedStaticSchedulingWithSetups.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2023 Vincent A. Cicirello
+ * Copyright (C) 2002-2026 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -390,9 +390,59 @@ public final class WeightedStaticSchedulingWithSetups
     PrintWriter out =
         new PrintWriter(
             new OutputStreamWriter(new FileOutputStream(filename), StandardCharsets.UTF_8), true);
-    WeightedStaticSchedulingWithSetupsWriter instanceWriter =
-        new WeightedStaticSchedulingWithSetupsWriter(this);
-    instanceWriter.toFile(out, instanceNumber);
+    toFile(out, instanceNumber);
     out.close();
+  }
+
+  /**
+   * Outputs a description of the instance data in the format based on that described in:
+   *
+   * <p>Vincent A. Cicirello. <a
+   * href="https://www.cicirello.org/publications/cicirello2003cmu.html">Weighted Tardiness
+   * Scheduling with Sequence-Dependent Setups: A Benchmark Library</a>. Technical Report,
+   * Intelligent Coordination and Logistics Laboratory, Robotics Institute, Carnegie Mellon
+   * University, Pittsburgh, PA, February 2003.
+   *
+   * <p>The data as output by this method varies from that format in that it does not output the
+   * "Generator Parameters" section. Instead, it has the "Begin Generator Parameters" and the "End
+   * Generator Parameters" block markers, but an empty block. The constructor of this class that
+   * takes an instance data file as input can correctly parse both the original and this modified
+   * format.
+   *
+   * @param out the output stream.
+   * @param instanceNumber An id for the problem instance.
+   */
+  public void toFile(PrintWriter out, int instanceNumber) {
+    out.print("Problem Instance: ");
+    out.println(instanceNumber);
+    out.print("Problem Size: ");
+    int n = numberOfJobs();
+    out.println(n);
+    out.println("Begin Generator Parameters");
+    out.println("End Generator Parameters");
+    out.println("Begin Problem Specification");
+    out.println("Process Times:");
+    for (int i = 0; i < n; i++) {
+      out.println(getProcessingTime(i));
+    }
+    out.println("Weights:");
+    for (int i = 0; i < n; i++) {
+      out.println(getWeight(i));
+    }
+    out.println("Duedates:");
+    for (int i = 0; i < n; i++) {
+      out.println(getDueDate(i));
+    }
+    out.println("Setup Times:");
+    for (int i = 0; i < n; i++) {
+      for (int j = 0; j < n; j++) {
+        if (i == j) {
+          out.printf("%d %d %d%n", -1, j, getSetupTime(i, j));
+        } else {
+          out.printf("%d %d %d%n", i, j, getSetupTime(i, j));
+        }
+      }
+    }
+    out.println("End Problem Specification");
   }
 }
